Fix build-blocking syntax issues
These are actual C-code and preprocessor macro changes, which are
required to get the code to build.

This doesn't include the changes to config.h, unixconf.h, and the
Makefiles, since editing those is a normal part of the build process
for old versions of NetHack.  I may make some of those changes in a
future commit, if I feel they are needed for everyone and not just
for my particular setup, but for now I left them.