Use Brass-specific options file, rather than the vanilla one.
I want this for me, for right now, because Brass currently doesn't
support a lot of the options in my .nethackrc file, and so it was
warning me about each and every one of them every time I ran the game
binary, meaning I had to hit space fifty-some times, at the end of
which it would ask me if I want it to randomly generate a character,
and space would be yes, making it VERY annoying to start a character
of a specific role or whatever.

Also, add a TODO list, purely for my own reference.