Go ahead and build nhdat even though the fakewiz level files are not generated.
I don't know if this is a game-breaking, ascension-preventing bug, or
a normal expected change specific to Brass that somehow got left out
of the makefiles.  (Did the Brass devs work on a platform that doesn't
use make, perhaps?)  That will need to be investigated; but for now,
this will allow the build to complete.  I've been doing it in my build
script, but I'm pretty sure it's not my peculiar setup, but something
anyone attempting to build will need.