Two fixes, the first leading to the second.
First, make the furthest level reached in any given branch be
considered interesting by #overview, even if no interesting features
have been encountered.  This will result in listing Gnomish Mines and
their first level when someone goes down the stairs and immediately
back up.  It will also produce a reminder of how far you've been--in
each branch--after retreating for any reason, without the need to
manually add an annotation.

Second, #overview was suppressing the range of level numbers for
Sokoban because the author realized that the values were wrong.  The
record of the furthest level reached was incorrect for builds-up
branches, always sticking with the deepest level even though it was
the entrance.  The overview patch neglected to do the same suppression
for Vlad's Tower and the level range ("36 to 38" or similar) there was
wrong.  This fixes the furthest level reached problem and also fixes

There was a distressing amount of trial and error involved.  The
dungeon layout structures are not exactly easy to work with, and I
never managed to get builds_up() based on branch data to work
correctly.  Basing it on dungeon data works as intended provided the
branch has more than one level, but it will yield incorrect result if
we ever add a single-level branch reached via stairs up rather than
stairs down.

Cherry-picked from commit 29f9a8a33d9748d23c027e548f7a99049ff63854.

For NetHack Brass, I did not include the part about making Sokoban get
harder as you climb toward the prize, partly it is controversial with
players, but mostly because it is unrelated to the #overview feature,
and I am not attempting to bring Brass up to date with 3.6, just with
the major community patches that all the other variants have. -- NAE