Dungeoncolors v1.2  (Bilious patch 17)
Allows the user to change the colors of the dungeon features and traps, by setting options in the config file.
       This patch works in all window ports that use character graphics, and has been tested on the TTY, X11, and Qt ports.

Changing the colors can be done with putting the following lines* into the configuration file:

DUNGEONCOLOR= 0 7 7 7 7 7 7 7 7 7 \
7 7 7 3 3 3 3 6 2 7 \
7 7 7 7 3 3 7 7 11 7 \
4 4 6 1 3 3 3 3 6 7 \
4

TRAPCOLORS= 6 6 7 3 6 1 7 12 4 9 0 0 3 3 5 5 13 7 7 12 12 10

``DUNGEONCOLOR'' are in the same order as ``DUNGEON'', and
``TRAPCOLORS'' as ``TRAPS''

And here's another set of colors* for the dungeon, with unlit
corridors darker than lit, and white stairs, just like Slash'em.

DUNGEONCOLOR= 0 7 7 7 7 7 7 7 7 7 \
7 7 7 3 3 3 3 6 2 7 \
0 7 15 15 3 3 7 7 11 7 \
4 4 6 1 3 3 3 3 6 7 \
4

The colors are:
0 black
1 red
2 green
3 brown
4 blue
5 magenta
6 cyan
7 gray
8 "no color"
9 orange
10 light green
11 yellow
12 light blue
13 light magenta
14 light cyan
15 white

Note:  the list of terrains is different in NetHack Brass, as follows:
   rock, vert.wall,  horz.wall, NW crnr, NE crnr, SW crnr, SE crnr, crosswall,
   top T, lower T, left T, right T, doorway, vertopen, horzopen, vertclosed,
   hrzclosed, Fe bars, tree, deadtree, dark rm, lit room, corridor, lit corr,
   pool, lit pool, ice, lit ice, bog, lit bog, ground, lit ground,
   grass, lit grass, carpet, lit carpet, up stair, dn stair, up lddr, dn lddr,
   altar, grave, throne, sink, fountain, lava, vert dn dbrdg, horz dn dbrdg,
   v raised dbrdg, h raised dbrdg, air, cloud, underwater

Sample colors (per Jonadab):
DUNGEONCOLOR=  0  7  7  7  7  7  7  7 \
               7  7  7  7  7  3  3  3 \
               3  6  2  0  7 15  0  7 \
               4 12  6 14  4 12  3 11 \
               2 10  5 13 10  5 10  5 \
              11  0 11  6 14  1  9  9 \
               9  9 15  7  4

The list of traps appears to be unchanged from vanilla.